refactor: extract generic headless-Chromium stack into frappe.utils.chromium
The Chromium process manager, CDP websocket client, Page driver and the
binary download/setup helpers are not PDF-specific — they're a general
"spin up headless chrome and drive a page" toolkit (already reused by the
preview/screenshot code and usable for scraping).

Move them out of frappe.utils.pdf_generator (and the download helpers out of
frappe.utils.print_utils) into a new frappe.utils.chromium package, and rename
the misleadingly-named ChromePDFGenerator -> ChromiumManager. pdf_generator
now holds only the PDF orchestration (Browser, pdf_merge, update_page_no.js).

BREAKING: the old import paths (frappe.utils.pdf_generator.chrome_pdf_generator
/ cdp_connection / page) and the chromium download symbols on
frappe.utils.print_utils are removed; import from frappe.utils.chromium instead.
