+++ frappe/utils/pdf.py
@@ -166,8 +166,8 @@ def measure_time(func):
 
 @measure_time
 def get_chrome_pdf(print_format, html, options, output, pdf_generator=None):
+	from frappe.utils.chromium import ChromiumManager
 	from frappe.utils.pdf_generator.browser import Browser
-	from frappe.utils.pdf_generator.chrome_pdf_generator import ChromePDFGenerator
 	from frappe.utils.pdf_generator.pdf_merge import PDFTransformer
 
 	if pdf_generator != "chrome":
@@ -175,7 +175,7 @@ def get_chrome_pdf(print_format, html, options, output, pdf_generator=None):
 		return
 	# scrubbing url to expand url is not required as we have set url.
 	# also, planning to remove network requests anyway 🤞
-	generator = ChromePDFGenerator()
+	generator = ChromiumManager()
 	try:
 		browser = Browser(generator, print_format, html, options)
 		transformer = PDFTransformer(browser)
